# docs/swagger.yaml
basePath: /
definitions:
  category_api.CategoryGetAllByPaginationRequest:
    properties:
      limit:
        type: integer
      page:
        type: integer
    type: object
  helper.Pages:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
    type: object
  product_api.DeleteRequest:
    properties:
      id:
        type: integer
    type: object
  product_api.ProductGetAllByPaginationRequest:
    properties:
      limit:
        type: integer
      page:
        type: integer
    type: object
  product_api.ProductRequest:
    properties:
      brand:
        type: string
      categoryId:
        type: integer
      name:
        type: string
    type: object
  product_api.SearchRequest:
    properties:
      value:
        type: string
    type: object
  product_api.UpdateRequest:
    properties:
      brand:
        type: string
      categoryId:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Basket service api provides category-product-order.
  title: Gin Basket Service API main
  version: "1.0"
paths:
  /category/:
    post:
      consumes:
      - application/json
      parameters:
      - description: pagination
        in: body
        name: paginationRequest
        schema:
          $ref: '#/definitions/category_api.CategoryGetAllByPaginationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/helper.Pages'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Gets all categories with paginated result
      tags:
      - Category
  /category/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Category CSV File
        in: formData
        name: categoryCsv
        required: true
        type: file
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add category by CSV file
      tags:
      - Category
...
  /product/:
    post:
      consumes:
      - application/json
      parameters:
      - description: pagination
        in: body
        name: paginationRequest
        schema:
          $ref: '#/definitions/product_api.ProductGetAllByPaginationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/helper.Pages'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Gets all products with paginated result
      tags:
      - Product
  /product/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product Info
        in: body
        name: productRequest
        schema:
          $ref: '#/definitions/product_api.ProductRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add product
      tags:
      - Product
  /product/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Product Id
        in: body
        name: deleteRequest
        schema:
          $ref: '#/definitions/product_api.DeleteRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete product
      tags:
      - Product
  /product/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: Search Value
        in: body
        name: searchRequest
        schema:
          $ref: '#/definitions/product_api.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Gets searched products
      tags:
      - Product
  /product/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: Products Update Values
        in: body
        name: updateRequest
        schema:
          $ref: '#/definitions/product_api.UpdateRequest'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update product
      tags:
      - Product
...
schemes:
- http
swagger: "2.0"